+++ R/BayesianMetaAnalysis.R
@@ -1,4 +1,4 @@
-# Copyright 2021 Observational Health Data Sciences and Informatics
+# Copyright 2022 Observational Health Data Sciences and Informatics
 #
 # This file is part of EvidenceSynthesis
 #
@@ -23,7 +23,7 @@ cleanData <- function(data,
     column <- columns[i]
     if (any(is.infinite(data[, column]))) {
       if (grid) {
-        warn(paste("Estimate(s) with infinite log-likelihood detected. Removing before computing meta-analysis."))  
+        warn(paste("Estimate(s) with infinite log-likelihood detected. Removing before computing meta-analysis."))
       } else {
         warn(paste("Estimate(s) with infinite",
                    column,
@@ -33,7 +33,7 @@ cleanData <- function(data,
     }
     if (any(is.na(data[, column]))) {
       if (grid) {
-        warn(paste("Estimate(s) with NA log-likelihood detected. Removing before computing meta-analysis."))  
+        warn(paste("Estimate(s) with NA log-likelihood detected. Removing before computing meta-analysis."))
       } else {
         warn(paste("Estimate(s) with NA",
                    column,
@@ -43,7 +43,7 @@ cleanData <- function(data,
     }
     if (any(data[, column] > maxValues[i])) {
       if (grid) {
-        warn(paste("Estimate(s) with positive log-likelihood detected. Removing before computing meta-analysis."))  
+        warn(paste("Estimate(s) with positive log-likelihood detected. Removing before computing meta-analysis."))
       } else {
         warn(sprintf("Estimate(s) with extremely high %s (>%s) detected. Removing before computing meta-analysis.",
                      column,
@@ -53,7 +53,7 @@ cleanData <- function(data,
     }
     if (any(data[, column] < minValues[i])) {
       if (grid) {
-        warn(paste("Estimate(s) with extremely low log-likelihood detected. Removing before computing meta-analysis."))  
+        warn(paste("Estimate(s) with extremely low log-likelihood detected. Removing before computing meta-analysis."))
       } else {
         warn(sprintf("Estimate(s) with extremely low %s (<%s) detected. Removing before computing meta-analysis.",
                      column,
@@ -87,7 +87,7 @@ isRmdCheck <- function() {
 
 isUnitTest <- function() {
   return(tolower(Sys.getenv("TESTTHAT", "")) == "true")
-  
+
 }
 
 #' Compute a Bayesian random-effects meta-analysis
@@ -106,6 +106,7 @@ isUnitTest <- function() {
 #'                             and tau, respectively.
 #' @param alpha                The alpha (expected type I error) used for the credible intervals.
 #' @param seed                 The seed for the random number generator.
+#' @param robust               Whether or not to use a t-distribution model (default: FALSE)
 #'
 #' @seealso
 #' [approximateLikelihood], [computeFixedEffectMetaAnalysis]
@@ -136,7 +137,7 @@ isUnitTest <- function() {
 #' estimate
 #'
 #' # (Estimates in this example will vary due to the random simulation)
-#' 
+#'
 #' @export
 computeBayesianMetaAnalysis <- function(data,
                                         chainLength = 1100000,
@@ -144,7 +145,8 @@ computeBayesianMetaAnalysis <- function(data,
                                         subSampleFrequency = 100,
                                         priorSd = c(2, 0.5),
                                         alpha = 0.05,
-                                        seed = 1) {
+                                        seed = 1,
+                                        robust = FALSE) {
   if (!supportsJava8()) {
     inform("Java 8 or higher is required, but older version was found. Cannot compute estimate.")
     estimate <- data.frame(mu = as.numeric(NA),
@@ -169,8 +171,9 @@ computeBayesianMetaAnalysis <- function(data,
                  "Result may be unreliable"))
     chainLength <- 110000
     burnIn <- 10000
+    Sys.sleep(1) # To avoid CRAN message about CPU time being more than 2.5. times elapsed time
   }
-  
+
   # Determine type based on data structure:
   if ("logRr" %in% colnames(data)) {
     inform("Detected data following normal distribution")
@@ -212,16 +215,34 @@ computeBayesianMetaAnalysis <- function(data,
     }
     dataModel$finish()
   } else if (is.list(data) && !is.data.frame(data)) {
-    inform("Detected (pooled) patient-level data")
-    type <- "pooled"
-    dataModel <- rJava::.jnew("org.ohdsi.metaAnalysis.CoxDataModel")
-    for (i in 1:length(data)) {
-      dataModel$addLikelihoodData(as.integer(data[[i]]$stratumId),
-                                  as.integer(data[[i]]$y),
-                                  as.numeric(data[[i]]$time),
-                                  as.numeric(data[[i]]$x))
+    if ("stratumId" %in% names(data[[1]])) {
+      inform("Detected (pooled) patient-level data")
+      type <- "pooled"
+      dataModel <- rJava::.jnew("org.ohdsi.metaAnalysis.CoxDataModel")
+      for (i in 1:length(data)) {
+        dataModel$addLikelihoodData(as.integer(data[[i]]$stratumId),
+                                    as.integer(data[[i]]$y),
+                                    as.numeric(data[[i]]$time),
+                                    as.numeric(data[[i]]$x))
+      }
+      dataModel$finish()
+    } else if ("point" %in% names(data[[1]])) {
+      inform("Detected data following adaptive grid distribution")
+      type <- "grid"
+      dataModel <- rJava::.jnew("org.ohdsi.metaAnalysis.ExtendingEmpiricalDataModel")
+      for (i in 1:length(data)) {
+        cleanedData <- as.data.frame(data[[i]])
+        cleanedData$value <- cleanedData$value - max(cleanedData$value)
+        cleanedData <- cleanData(cleanedData,
+                                 c("point", "value"),
+                                 minValues = c(-100, -1e6),
+                                 maxValues = c(100, 0))
+        dataModel$addLikelihoodParameters(cleanedData$point, cleanedData$value)
+      }
+      dataModel$finish()
+    } else {
+      abort("Unknown input data format")
     }
-    dataModel$finish()
   } else {
     inform("Detected data following grid distribution")
     type <- "grid"
@@ -243,20 +264,34 @@ computeBayesianMetaAnalysis <- function(data,
     }
     dataModel$finish()
   }
-  
+
   inform("Performing MCMC. This may take a while")
-  
+
   prior <- rJava::.jnew("org.ohdsi.metaAnalysis.HalfNormalOnStdDevPrior", 0, as.numeric(priorSd[2]))
-  
-  metaAnalysis <- rJava::.jnew("org.ohdsi.metaAnalysis.Runner",
-                               rJava::.jcast(rJava::.jnew("org.ohdsi.metaAnalysis.MetaAnalysis",
-                                                          rJava::.jcast(dataModel, "org.ohdsi.metaAnalysis.DataModel"),
-                                                          rJava::.jcast(prior, "org.ohdsi.metaAnalysis.ScalePrior"),
-                                                          as.numeric(priorSd[1])), "org.ohdsi.metaAnalysis.Analysis"),
-                               as.integer(chainLength),
-                               as.integer(burnIn),
-                               as.integer(subSampleFrequency),
-                               as.numeric(seed))
+
+  if(robust){
+    metaAnalysis <- rJava::.jnew("org.ohdsi.mcmc.Runner",
+                                 rJava::.jcast(rJava::.jnew("org.ohdsi.metaAnalysis.RobustMetaAnalysis",
+                                                            rJava::.jcast(dataModel, "org.ohdsi.metaAnalysis.DataModel"),
+                                                            rJava::.jcast(prior, "org.ohdsi.metaAnalysis.ScalePrior"),
+                                                            as.numeric(priorSd[1])),
+                                               "org.ohdsi.mcmc.Analysis"),
+                                 as.integer(chainLength),
+                                 as.integer(burnIn),
+                                 as.integer(subSampleFrequency),
+                                 as.numeric(seed))
+  }else{
+    metaAnalysis <- rJava::.jnew("org.ohdsi.mcmc.Runner",
+                                 rJava::.jcast(rJava::.jnew("org.ohdsi.metaAnalysis.MetaAnalysis",
+                                                            rJava::.jcast(dataModel, "org.ohdsi.metaAnalysis.DataModel"),
+                                                            rJava::.jcast(prior, "org.ohdsi.metaAnalysis.ScalePrior"),
+                                                            as.numeric(priorSd[1])), "org.ohdsi.mcmc.Analysis"),
+                                 as.integer(chainLength),
+                                 as.integer(burnIn),
+                                 as.integer(subSampleFrequency),
+                                 as.numeric(seed))
+  }
+
   metaAnalysis$setConsoleWidth(getOption("width"))
   metaAnalysis$run()
   parameterNames <- metaAnalysis$getParameterNames()